Split save_data_utility into two classes.
