+++ rpcs3/rpcs3qt/save_data_dialog.cpp
@@ -2,7 +2,7 @@
 #include "Emu/Memory/Memory.h"
 
 #include "save_data_dialog.h"
-#include "save_data_utility.h"
+#include "save_data_list_dialog.h"
 
 s32 save_data_dialog::ShowSaveDataList(std::vector<SaveDataEntry>& save_entries, s32 focused, bool isSaving, vm::ptr<CellSaveDataListSet> listSet)
 {
