+++ rpcs3/rpcs3qt/save_data_info_dialog.h
@@ -1,0 +1,24 @@
+#pragma once
+
+// I just want the struct for the save data.
+#include "stdafx.h"
+#include "Emu/Memory/Memory.h"
+#include "Emu/Cell/Modules/cellSaveData.h"
+
+#include <QDialog>
+#include <QTableWidget>
+
+//Used to display the information of a savedata.
+//Not sure about what information should be displayed.
+class save_data_info_dialog :public QDialog
+{
+	Q_OBJECT
+
+public:
+	explicit save_data_info_dialog(const SaveDataEntry& save, QWidget* parent = nullptr);
+private:
+	void UpdateData();
+
+	SaveDataEntry m_entry;
+	QTableWidget* m_list;
+};
